302213 - Implement ids/Users endpoint and move common query to a helper
/ids/Users endpoint maps user ids to usernames and is the same as the
users /:id query endpoint but with a limited filter subset.

Moving the user querying from the UsersController to a UserHelper helper
that is imported to both UsersController (for current users querying)
and the RootController for the new /ids/Users endpoint.

Also updated the ActiveRecordQueryBuilder to be able to provide a list
of "allowed" operators / fields for filtering.
