+++ controllers/users_controller.rb
@@ -1,8 +1,5 @@
 class UsersController < ApplicationController
-
-  VALID_ATTRIBUTES = %W{id name userName emails groups externalId}.collect(&:downcase)
-  MAX_ITEMS_PER_PAGE = 1000
-  DEFAULT_ITEMS_PER_PAGE = 100
+  helpers UserHelper
 
   # Create a User
   # https://github.com/cloudfoundry/uaa/blob/master/docs/UAA-APIs.rst#create-a-user-post-users
@@ -99,31 +96,9 @@ class UsersController < ApplicationController
       # This error doesn't show up in logs
       raise Aok::Errors::ScimFilterError.new($!.message)
     end
-    filter = true if filter.blank? || filter =='""' #XXX bug in scim-query-filter-parser-rb
-    start_index = params[:startIndex] || 1
-    start_index = [start_index.to_i, 1].max
-    items_per_page = params[:count] || DEFAULT_ITEMS_PER_PAGE
-    items_per_page = [items_per_page.to_i, MAX_ITEMS_PER_PAGE].min
-    identities = Identity.
-      where(filter).
-      includes(:groups =>[:parent_groups]).
-      limit(items_per_page).
-      offset(start_index - 1)
-    resources = []
+    users = query_users_as_scim(filter, params)
 
-    identities.each_with_index do |identity, index|
-      resources.push(user_hash(identity, attributes))
-    end
-
-    response = {
-      'totalResults' => Identity.where(filter).count,
-      'itemsPerPage' => items_per_page,
-      'startIndex' => start_index,
-      'schemas' => ["urn:scim:schemas:core:1.0"],
-      'resources' => resources,
-    }
-
-    response.to_json
+    users.to_json
   end
 
   # Delete a User
@@ -152,47 +127,6 @@ class UsersController < ApplicationController
       user.email = user_details.emails.first['value']
     end
     return user
-  end
-
-  def attributes
-    return nil unless params[:attributes]
-    attrs = params[:attributes].downcase.split(',')
-    return attrs & VALID_ATTRIBUTES
-  end
-
-  def user_hash user, attrs = nil
-    attrs ||= VALID_ATTRIBUTES
-    user_data = {
-      'schemas' => ['urn:scim:schemas:core:1.0']
-    }
-    user_data['externalId'] = user.username if attrs.include?('externalid')
-    user_data['id'] = user.guid if attrs.include?('id')
-    user_data['meta'] = {
-      'version' => user.version,
-      'created' => user.created_at.utc.strftime(UAA_DATE_FORMAT),
-      'lastModified' => user.updated_at.utc.strftime(UAA_DATE_FORMAT),
-    }
-
-    f = user.family_name
-    g = user.given_name
-    if (f or g) && attrs.include?('name')
-      n = user_data['name'] = {}
-      n['familyName'] = f if f
-      n['givenName'] = g if g
-    end
-    user_data['userName'] = user.username if attrs.include?('username')
-    user_data["emails"] = [
-      {"value" => user.email}
-    ] if attrs.include?('emails')
-    if attrs.include?('groups')
-      user_data['groups'] = user.ascendant_groups.collect do |group|
-        {
-          'display' => group.name,
-          'value' => group.guid
-        }
-      end
-    end
-    return user_data
   end
 
   def scim_user_response user
